# Row writer optimization for bulk insert