--- a/docker/demo/config/hoodie-schema.avsc
+++ b/docker/demo/config/hoodie-schema.avsc
@@ -88,7 +88,7 @@
                 "name": "abc",
                 "size": 5,
                 "logicalType": "decimal",
-                "precision": 10,
+                "precision": 12,
                 "scale": 6
             }
         },