--- a/hudi-utilities/src/test/resources/streamer-config/source_evolved.avsc
+++ b/hudi-utilities/src/test/resources/streamer-config/source_evolved.avsc
@@ -93,7 +93,7 @@
         "name": "abc",
         "size": 5,
         "logicalType": "decimal",
-        "precision": 10,
+        "precision": 12,
         "scale": 6
       }
     },