--- a/hudi-utilities/src/test/resources/streamer-config/target.avsc
+++ b/hudi-utilities/src/test/resources/streamer-config/target.avsc
@@ -104,7 +104,7 @@
         "name": "abc",
         "size": 5,
         "logicalType": "decimal",
-        "precision": 10,
+        "precision": 12,
         "scale": 6
       }
     },