### Change Logs

We found inefficiencies with bulk insert use case where in delta streamer code path we were taking a performance hit for doing row to avro conversion when transformer is hooked on. This PR addresses to avoid these unnecessary conversions by directly passing hoodie spark record to the write client. 

### Impact

Initial testing of this PR shows the following benefits

master(without changes):
1550.62s user 268.99s system 496% cpu 6:06.20 total
1548.00s user 214.47s system 500% cpu 5:52.47 total
1530.54s user 205.70s system 507% cpu 5:42.22 total

row(with changes):
1398.94s user 16.19s system 560% cpu 4:12.48 total
1387.49s user 16.15s system 560% cpu 4:10.62 total
1386.58s user 39.09s system 540% cpu 4:23.62 total

### Risk level (write none, low medium or high below)

low, we are trying to enhance this performance boost only for row writer use case.

### Contributor's checklist

- [ ] Read through [contributor's guide](https://hudi.apache.org/contribute/how-to-contribute)
- [ ] Change Logs and Impact were stated clearly
- [ ] Adequate tests were added if applicable
- [ ] CI passed